requirements.runtime.txt
numpy>=1.21.0,<2.0.0
pyspark==3.3.0
pandas>=1.3.0,<2.0.0
findspark
Pillow>=9.0.0
scikit-learn>=1.0.0,<1.6.0
torch==2.0.1
torchvision==0.15.2